+++ hslua-classes/test/HsLua/ClassesTests.hs
@@ -1,33 +1,24 @@
-{-
-Copyright © 2017-2021 Albert Krewinkel
+{-|
+Module      : HsLua.ClassesTests
+Copyright   : © 2007–2012 Gracjan Polak;
+              © 2012–2016 Ömer Sinan Ağacan;
+              © 2017-2021 Albert Krewinkel
+License     : MIT
+Maintainer  : Albert Krewinkel <tarleb+hslua@zeitkraut.de>
+Stability   : beta
+Portability : non-portable (depends on GHC)
 
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
+Test that conversions from and to the Lua stack are isomorphisms.
 -}
-{-| Test that conversions from and to the lua stack are isomorphisms.
--}
-module HsLua.TypesTests (tests) where
+module HsLua.ClassesTests (tests) where
 
 import Control.Monad (forM, forM_)
 import Data.ByteString (ByteString)
 import Data.Map (Map)
 import Data.Set (Set)
 import HsLua as Lua
+import HsLua.Class.Peekable
+import HsLua.Class.Pushable
 import Lua.Arbitrary ()
 import Test.QuickCheck
 import Test.QuickCheck.Instances ()